--- playwright-e2e/data/fake-data-utils.ts
@@ -1,5 +1,0 @@
-import { faker } from '@faker-js/faker';
-
-export const generateUserName = (namePrefix: string): string => {
-  return `${namePrefix}-${faker.name.lastName()}`;
-};
